Move log.sql tests to a log package
So that we can group all log tests (log.sql, log.bind-parameters) in the
same package.
