+++ extensions/hibernate-orm/deployment/src/test/java/io/quarkus/hibernate/orm/log/LogSqlFormatSqlDefaultValueTest.java
@@ -1,4 +1,4 @@
-package io.quarkus.hibernate.orm.logsql;
+package io.quarkus.hibernate.orm.log;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
